Extract auto-research evidence packet helpers
